+++ dispatch-c/src/verify.h
@@ -1,10 +1,9 @@
-#include "integer.h"
 #include <vector>
 using namespace std;
 
-bool judgeSingleNode(vector<integer>& arrayList, int i, int n, int t);
+bool judgeSingleNode(vector<uint64_t>& arrayList, int i, int n, int t);
 
-void judge(vector<integer>& arrayList, vector<int>& countMap, int n, int t);
+void judge(vector<uint64_t>& arrayList, vector<int>& countMap, int n, int t);
 
-double sampling_verify(vector<integer>& arrayList, int samplingNum, int n, int t);
+double sampling_verify(vector<uint64_t>& arrayList, int samplingNum, int n, int t);
